move: add dto/user package and move dto about user
