+++ src/main/java/EFUB/homepage/controller/PassController.java
@@ -1,7 +1,7 @@
 package EFUB.homepage.controller;
 
 import EFUB.homepage.dto.pass.PassDto;
-import EFUB.homepage.dto.pass.PassResDto;
+import EFUB.homepage.dto.pass.PassResultResDto;
 import EFUB.homepage.service.PassService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -17,12 +17,12 @@ public class PassController {
 	private final PassService passService;
 
 	@PostMapping("/mid")
-	public PassResDto checkMidPass(@RequestBody PassDto passDto) {
+	public PassResultResDto checkMidPass(@RequestBody PassDto passDto) {
 		return passService.checkMidPass(passDto);
 	}
 
 	@PostMapping("/final")
-	public PassResDto checkFinPass(@RequestBody PassDto passDto) {
+	public PassResultResDto checkFinPass(@RequestBody PassDto passDto) {
 		return passService.checkFinPass(passDto);
 	}
 }
