--- src/main/java/EFUB/homepage/dto/common/PassReqDto.java
@@ -1,12 +1,0 @@
-package EFUB.homepage.dto.common;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-
-@Getter
-@NoArgsConstructor
-public class PassReqDto {
-	private ArrayList<Long> users;
-}
