+++ src/main/java/EFUB/homepage/dto/pass/PassResDto.java
@@ -1,16 +1,21 @@
 package EFUB.homepage.dto.pass;
 
+import EFUB.homepage.dto.user.UserResDto;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
+@Getter
 @NoArgsConstructor
-@Data
 public class PassResDto {
-	private String result;
+	private Long count;
+	private List<UserResDto> users;
 
 	@Builder
-	public PassResDto(String result) {
-		this.result = result;
+	public PassResDto(Long count, List<UserResDto> users) {
+		this.count = count;
+		this.users = users;
 	}
 }
