+++ src/main/java/EFUB/homepage/dto/pass/PassResultResDto.java
@@ -1,0 +1,16 @@
+package EFUB.homepage.dto.pass;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@Data
+public class PassResultResDto {
+	private String result;
+
+	@Builder
+	public PassResultResDto(String result) {
+		this.result = result;
+	}
+}
